refactor: add directories to sdk
