+++ batcher/aligned-batcher/src/types/batch_queue.rs
@@ -5,7 +5,7 @@ use futures_util::stream::SplitSink;
 use tokio::{net::TcpStream, sync::RwLock};
 use tokio_tungstenite::{tungstenite::Message, WebSocketStream};
 
-use aligned_sdk::types::{VerificationData, VerificationDataCommitment};
+use aligned_sdk::core::types::{VerificationData, VerificationDataCommitment};
 
 pub(crate) type BatchQueueEntry = (
     VerificationData,
