+++ batcher/aligned-batcher/src/zk_utils/mod.rs
@@ -3,7 +3,7 @@ use crate::halo2::ipa::verify_halo2_ipa;
 use crate::halo2::kzg::verify_halo2_kzg;
 use crate::risc_zero::verify_risc_zero_proof;
 use crate::sp1::verify_sp1_proof;
-use aligned_sdk::types::{ProvingSystemId, VerificationData};
+use aligned_sdk::core::types::{ProvingSystemId, VerificationData};
 use log::{debug, warn};
 
 pub(crate) fn verify(verification_data: &VerificationData) -> bool {
