+++ batcher/aligned-sdk/src/core/errors.rs
