+++ batcher/aligned-sdk/src/lib.rs
@@ -1,8 +1,15 @@
-mod batch;
-mod communication;
-pub mod errors;
-mod eth;
-pub mod protocol;
+pub mod core {
+    pub mod errors;
+    pub mod types;
+    pub mod utils;
+}
+
+pub mod communication {
+    mod batch;
+    pub(crate) mod messaging;
+    pub mod protocol;
+}
+
+pub mod eth;
+
 pub mod sdk;
-pub mod types;
-pub mod utils;
